refactored all the test data into testingData.tsx
